Update INI to remove saved [MainWindow] position
The shipping Repo should not include the [MainWindow] position. It may be off monitor for some.
In the future I would have 2 INI files so a separate one can save Window Positions that are user specific.
Or maybe this INI should NOT be in the Repo at all? Need to test without it.