Fix bug FnSplit LFN fails
FileExtension() failed with LFN because FNInfoType GROUP had Name varible defined as CSTRING(9) i.e. old 8.3 limit.
This caused the call to FnSplit() to not work to return the correct .EXT.

FileExtension() returned a String Slice [] with end as LEN(Extension) that would be an invalid slice if not .EXT so would GPF. Change to SUB().