Fix Window code to show up to 128 rules, was
There was a limit of 18 rules/dll on the window with 17 already defined, and 32 maximum.
New Maximum 128 for both, but probably 72/dll is reasonable max.
The default window size shows 18, added message to warn window needs to be wider.

With monitors much bigger should resize Wizard window to be taller and wider, maybe remove BMP.